HDF5 integration
The main object of this commit is to integrate a result file, which stores embeddings and other information. This is done using hdf5, because it is interoperable and fast (depending on the implementation). For an overview of what is in the file, see docembedder/datamodel.py. The class is setup in a way that is should be possible to create other data models, but I don't expect we'll actually need that.

A new system for defining the windows/setup of with which patents the models are trained. The idea is that each of the HDF5 files will have the same system applied, so that different models can be compared, because they are trained on the same patents. This resides in docembedder/utils.py.

Updated benchmarks to showcase the new HDF5 file system, see the examples folder. They are not meant to be the final benchmarks.

Moved the models to a separate subpackage, since the number of modules has been going up.

Started a new Analysis class in docembedder/analysis2.py that works with the new HDF5 file system. Obviously, in the end it will replace analysis.py and be renamed.

Added parallel computation of the models.